Client & ClientConfig handle input validation
Config validation was happening in the Client struct, and in some cases, was required by the code calling the client. The config struct has been split away from the Client, and handles it's validation on it's own. The client simply accepts the Config struct now, relying on that config struct to have handled the necessary validation